seeds and migrations working

+++ api/migrations/20231226190932_migration_name.js
@@ -10,7 +10,7 @@ exports.up = function(knex) {
     table.date('date').notNullable();
     table.time('startTime').notNullable();
     table.time('endTime').notNullable();
-    table.integer('totalHours').notNullable()
+    table.decimal('totalHours').notNullable()
   })
 };
